Edit player names update after any text change > home screen FAB removed
