+++ app/src/main/java/com/example/dotsboxes/Fragments/GameFragment.java
@@ -1,4 +1,5 @@
 package com.example.dotsboxes.Fragments;
+import android.app.Dialog;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -6,9 +7,13 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.cardview.widget.CardView;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.WindowManager;
 import android.view.animation.Animation;
 import android.view.animation.LinearInterpolator;
 import android.view.animation.TranslateAnimation;
@@ -43,12 +48,16 @@ public class GameFragment extends Fragment {
         setPlayerColors(cvP1, cvP2, p1Name, p2Name);
         setPlayerAvatars(ivP1, ivP2);
 
+        binding.btnHelp.setOnClickListener(view -> showHelpDialog());
+
+        binding.btnSettings.setOnClickListener(view -> replaceFragment(new SettingsFragment()));
+
         gameView.setUpReferences(p1Score, p2Score, p1Name, p2Name, statusDisplay, btnPlayAgain, p1Turn, p2Turn);
         return myView;
     }
 
     private void setPlayerAvatars(ImageView ivP1, ImageView ivP2) {
-        SharedPreferences sharedPreferences = MainActivity.getContext().getSharedPreferences(PrefUtility.SHARED_PREF_NAME, Context.MODE_PRIVATE);
+        SharedPreferences sharedPreferences = getContext().getSharedPreferences(PrefUtility.SHARED_PREF_NAME, Context.MODE_PRIVATE);
         boolean playComputer = sharedPreferences.getBoolean(PrefUtility.IS_PLAY_COMPUTER, false);
 
         String playerAvatar1 = sharedPreferences.getString(PrefUtility.PLAYER_AVATAR_1, PrefUtility.DEFAULT_PLAYER_AVATAR_1);
@@ -60,7 +69,7 @@ public class GameFragment extends Fragment {
     }
 
     private void setPlayerColors(CardView cvP1, CardView cvP2, TextView tvP1, TextView tvP2) {
-        SharedPreferences sharedPreferences = MainActivity.getContext().getSharedPreferences(PrefUtility.SHARED_PREF_NAME, Context.MODE_PRIVATE);
+        SharedPreferences sharedPreferences = getContext().getSharedPreferences(PrefUtility.SHARED_PREF_NAME, Context.MODE_PRIVATE);
         String playerColor1 = sharedPreferences.getString(PrefUtility.PLAYER_COLOR_1, "");
         String playerColor2 = sharedPreferences.getString(PrefUtility.PLAYER_COLOR_2, "");
 
@@ -68,6 +77,23 @@ public class GameFragment extends Fragment {
         cvP2.setCardBackgroundColor(getResources().getColor(PrefUtility.getColor(playerColor2)));
         tvP1.setTextColor(getResources().getColor(PrefUtility.getColor(playerColor1)));
         tvP2.setTextColor(getResources().getColor(PrefUtility.getColor(playerColor2)));
+    }
+
+    private void showHelpDialog() {
+        Dialog dialog = new Dialog(getContext());
+        WindowManager.LayoutParams lp = new WindowManager.LayoutParams();
+        dialog.setContentView(R.layout.dialog_help);
+        lp.copyFrom(dialog.getWindow().getAttributes());
+        lp.width = WindowManager.LayoutParams.MATCH_PARENT;
+        lp.height = WindowManager.LayoutParams.MATCH_PARENT;
+        dialog.show(); dialog.getWindow().setAttributes(lp);
+        dialog.findViewById(R.id.close_corner).setOnClickListener(view -> dialog.dismiss());
+    }
+
+    private void replaceFragment(Fragment fragment) {
+        FragmentManager fragmentManager = getParentFragmentManager();
+        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+        fragmentTransaction.replace(R.id.frame_layout, fragment).setReorderingAllowed(true).addToBackStack(null).commit();
     }
 
     public static void animateTurn(ImageView visiblePlayer, ImageView invisiblePlayer) {
