+++ app/src/main/java/com/example/dotsboxes/MainActivity.java
@@ -4,7 +4,6 @@ import androidx.appcompat.app.AppCompatDelegate;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-import android.app.Dialog;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -12,7 +11,6 @@ import android.os.Bundle;
 import android.util.DisplayMetrics;
 import android.view.WindowManager;
 import com.example.dotsboxes.Fragments.GameTypeFragment;
-import com.example.dotsboxes.Fragments.SettingsFragment;
 import com.example.dotsboxes.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
@@ -22,15 +20,11 @@ public class MainActivity extends AppCompatActivity {
     private static boolean isFirstTime;
     private SharedPreferences sharedPreferences;
     private SharedPreferences.Editor editor;
-    private final SettingsFragment settingsFragment = new SettingsFragment();
-
-    private static Context context;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-        MainActivity.context = getApplicationContext();
         hideActionBar();
         setDeviceDimensions();
         fetchStoredData();
@@ -44,11 +38,6 @@ public class MainActivity extends AppCompatActivity {
         }
 
         replaceFragment(new GameTypeFragment());
-
-        binding.btnHelp.setOnClickListener(view -> showHelpDialog());
-
-        binding.btnSettings.setOnClickListener(view -> replaceFragment(settingsFragment));
-
     }
 
     private void setDefaultPrefs() {
@@ -73,17 +62,6 @@ public class MainActivity extends AppCompatActivity {
         fragmentTransaction.replace(R.id.frame_layout, fragment).setReorderingAllowed(true).addToBackStack(null).commit();
     }
 
-    private void showHelpDialog() {
-        Dialog dialog = new Dialog(this);
-        WindowManager.LayoutParams lp = new WindowManager.LayoutParams();
-        dialog.setContentView(R.layout.dialog_help);
-        lp.copyFrom(dialog.getWindow().getAttributes());
-        lp.width = WindowManager.LayoutParams.MATCH_PARENT;
-        lp.height = WindowManager.LayoutParams.MATCH_PARENT;
-        dialog.show(); dialog.getWindow().setAttributes(lp);
-        dialog.findViewById(R.id.close_corner).setOnClickListener(view -> dialog.dismiss());
-    }
-
     private void setDeviceDimensions() {
         DisplayMetrics displayMetrics = new DisplayMetrics();
         getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
@@ -100,7 +78,5 @@ public class MainActivity extends AppCompatActivity {
         editor = sharedPreferences.edit();
         isFirstTime = sharedPreferences.getBoolean(PrefUtility.IS_FIRST_TIME, true);
     }
-
-    public static Context getContext() { return MainActivity.context; }
 
 }
